+++ src/main/java/com/softserveinc/ita/mmakoviichuk/models/TitleData.java
@@ -1,4 +1,4 @@
-package com.softserveinc.ita.models;
+package com.softserveinc.ita.mmakoviichuk.models;
 
 import lombok.Builder;
 import lombok.Data;
